create utilities function folder and api.js
